Movement reworked
Movement changed from an interrupt based event  driven functionality to
more of a polling based system.  The keydown and keyup events modify a
map on the keyboard manager, and then on clock tick, the canvas wrapper
checks against this map to see what needs to be done.  End result is
smoother controls and better handling for multiple key presses.